// src/shared/services/redis/redis.type.ts
export enum REDIS_TYPE {
  REDIS_SUBSCRIBER_CLIENT = "REDIS_SUBSCRIBER_CLIENT",
  REDIS_PUBLISHER_CLIENT = "REDIS_PUBLISHER_CLIENT",

  REDIS_SOCKET_EVENT_PING = "REDIS_SOCKET_EVENT_PING",
  REDIS_SOCKET_EVENT_SEND_NAME = "REDIS_SOCKET_EVENT_SEND_NAME",
  REDIS_SOCKET_EVENT_EMIT_ALL_NAME = "REDIS_SOCKET_EVENT_EMIT_ALL_NAME",
  REDIS_SOCKET_EVENT_EMIT_AUTHENTICATED_NAME = "REDIS_SOCKET_EVENT_EMIT_AUTHENTICATED_NAME",
  REDIS_SOCKET_SUBSCRIBE_EVENT_QUIZ_CHANGE = "REDIS_SOCKET_SUBSCRIBE_EVENT_QUIZ_CHANGE",
}